@@ -1,26 +1,21 @@
-name: flutter_google_places
-description: Google places autocomplete widgets for flutter. No wrapper, use https://pub.dev/packages/google_maps_webservice
-version: 0.4.0
-repository: https://github.com/fluttercommunity/flutter_google_places
+name: neom_google_places
+description: Google places autocomplete widgets for flutter maintained by Cyberneom.
+version: 1.0.1
+repository: https://github.com/Cyberneom/neom_google_places
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.17.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
-  flutter:
-    sdk: flutter
-  google_api_headers: ^1.3.0
-  google_maps_webservice: ^0.0.20-nullsafety.5
-  http: ^0.13.4
-  rxdart: ^0.27.5
+  google_api_headers: ^4.1.2
+  neom_maps_services: ^1.0.3
+  http: ^1.2.1
+  rxdart: ^0.28.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  lint: ^1.10.0
+  flutter_lints: ^4.0.0
 
 flutter:
   assets:
-    - packages/flutter_google_places/assets/google_black.png
-    - packages/flutter_google_places/assets/google_white.png
+    - packages/neom_google_places/assets/google_black.png
+    - packages/neom_google_places/assets/google_white.png
