@@ -1,5 +1,5 @@
-library flutter_google_places;
+library neom_google_places;
 
-export 'src/flutter_google_places.dart';
+export 'src/neom_google_places.dart';
 export 'src/places_autocomplete_field.dart';
 export 'src/places_autocomplete_form_field.dart';
